All reward tests passed for 1a, 2a, 3a. Might need refactoring.

+++ src/main/java/software/aoc/day02/a/Range.java
@@ -1,4 +1,4 @@
 package software.aoc.day02.a;
 
-public record Range(int start, int end) {
+public record Range(long start, long end) {
 }
